+++ springboot-basic/src/main/java/com/example/commandlineapplication/global/io/Input.java
@@ -1,11 +1,11 @@
-package com.example.commandlineapplication.io;
+package com.example.commandlineapplication.global.io;
 
 import java.util.Scanner;
 
 public class Input {
   private static final Scanner scanner = new Scanner(System.in);
 
-  public String input() {
+  public String selectMenu() {
     return scanner.nextLine();
   }
 }
