+++ springboot-basic/src/main/java/com/example/commandlineapplication/global/io/Output.java
@@ -1,4 +1,4 @@
-package com.example.commandlineapplication.io;
+package com.example.commandlineapplication.global.io;
 
 public class Output {
   private static final String menu = "=== Voucher Program ===\nType exit to exit the program.\nType create to create a new voucher.\nType list to list all vouchers.";
